requirements.txt
Django==4.0.3
djangorestframework==3.13.1
pytest==7.1.1
